% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{dsc_null_test}
\alias{dsc_null_test}
\title{Test for zero discrimination (DSC=0)}
\usage{
dsc_null_test(X, Y, S, V, Spp, vcov_estimator = sandwich::vcovHAC)
}
\arguments{
\item{X}{A numeric vector representing the forecast values.}

\item{Y}{A numeric vector representing the actual realization values.}

\item{S}{Scoring function.}

\item{V}{Identification function, first derivative of the score (V := S').}

\item{Spp}{Second derivative of the score (S'').}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}
}
\value{
A tibble containing the results of the dynamic stochastic calibration tests, including
        the DSC value, DSC variance, DSC p-value, Mincer-Zarnowitz p-value, and Wald test p-value.
}
\description{
This function implements tests for the null hypothesis of DSC=0, using
forecast (X) and realization (Y) data. It employs the generalized Chi-square
test of the score decomposition. Also the Wald and F test pendants are
reported.
}
\details{
The function first regresses Y on X to extract fitted values and residuals,
         then calculates the variance-covariance matrix using the HAC estimator.
         It proceeds to compute the DSC value, its variance, and p-value using the
         generalized Chi-square asymptotics.

         Note: The function requires the `sandwich`, `MASS`, `car`, and `CompQuadForm`
         packages for various calculations.
}
\examples{
# Example usage:
S <- function(x, y) (x - y)^2
V <- function(x, y) 2*(x - y)
Spp <- function(x, y) 2
set.seed(1234)
X <- rnorm(100)
Y <- X + rnorm(100)
test_results <- dsc_null_test(X, Y, S, V, Spp)
print(test_results)

}
