% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{get_pval}
\alias{get_pval}
\title{Calculate p-values based on specified theory.}
\usage{
get_pval(dt, tt)
}
\arguments{
\item{dt}{An asy_vars object from function .}

\item{tt}{Number of observations.}
}
\value{
A data frame containing p-values for different components: s_pval, s_pvalDM, mcb_pval, and dsc_pval.
}
\description{
This function calculates p-values based on the provided theory for different components.
It takes an asy_vars object containing relevant variables and the number of observations.
}
\details{
This function computes p-values for s, mcb, and dsc components based on the specified theory.
It also calculates the original DM p-value for the s component.
}
