% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_hac_helperfuns.R
\name{hac_qr_galvao_yoon_internal}
\alias{hac_qr_galvao_yoon_internal}
\title{Galvão–Yoon (2024) HAC helper functions for time-series quantile regression}
\description{
Internal helper functions used to compute HAC covariance matrices for quantile regression
following Galvao and Yoon (2024).  The functions below are adapted from the supplementary material / code of
 Galvao, A. F., & Yoon, J. (2024), JASA 119(547), 2305–2316.
 They are included here for internal use only.
}
\details{
Source: Galvao, A. F., & Yoon, J. (2024). HAC Covariance Matrix Estimation in Quantile Regression.
\emph{Journal of the American Statistical Association}, 119(547), 2305–2316.
\doi{10.1080/01621459.2023.2257365}
}
\keyword{internal}
