% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{mcb_null_test}
\alias{mcb_null_test}
\title{Test for zero Miscalibration (MCB=0)}
\usage{
mcb_null_test(X, Y, S, V, Spp, vcov_estimator = sandwich::vcovHAC)
}
\arguments{
\item{X}{A numeric vector representing the forecast values.}

\item{Y}{A numeric vector representing the actual realization values.}

\item{S}{scoring function}

\item{V}{identification function, first derivative of score V:= S'}

\item{Spp}{second derviative of score S''}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}
}
\value{
A tibble containing the results of the mis-calibration tests, including
        the MCB value, MCB variance, MCB p-value, Mincer-Zarnowitz p-value,
        and F-test p-value.
}
\description{
This function implements tests for the null hypothesis of zero miscalibration (MCB=0)
using forecast (X) and realization (Y) data. It employs two methods for testing:
a generalized Chi-square test of the loss function decomposition as proposed in the accombined paper
and the classical Mincer-Zarnowitz test.
}
\details{
The function first regresses Y on X to extract fitted values and residuals,
         then calculates the variance-covariance matrix using a HAC estimator.
         It proceeds to compute the MCB value, its variance, and p-value using the
         generalized Chi-square asymptotics. The Mincer-Zarnowitz test is conducted
         as a comparative methodology. The F-test p-value is also computed for
         assessing the joint hypothesis that the intercept is zero and the slope is one
         in the linear regression of Y on X.

         Note: The function requires the `sandwich`, `MASS`, `car`, and `CompQuadForm`
         packages for various calculations.
}
\examples{
# Example usage:
S <- function(x, y) (x - y)^2
V <- function(x, y) 2*(x - y)
Spp <- function(x, y) 2
set.seed(1234)
X <- rnorm(100)
Y <- X + rnorm(100)
test_results <- mcb_null_test(X, Y, S, V, Spp)
print(test_results)

}
