% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{MZ.test.mean}
\alias{MZ.test.mean}
\title{Mincer-Zarnowitz Mean Test}
\usage{
MZ.test.mean(haty, y, vcov_estimator = sandwich::NeweyWest)
}
\arguments{
\item{haty}{Numeric vector of forecasted values.}

\item{y}{Numeric vector of actual values.}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}
}
\value{
A list containing several components:
  - `estimate`: Estimated coefficients from the Mincer-Zarnowitz regression.
  - `CI`: Confidence intervals for the estimated coefficients.
  - `WaldStat`: Wald statistic for testing the joint hypothesis that the intercept is zero and the slope is one.
  - `Wald.pval`: p-value associated with the Wald statistic.
  - `fittedvalues`: A tibble with the forecasted values and the corresponding fitted values from the regression.
}
\description{
This function performs the classical Mincer-Zarnowitz test to evaluate the accuracy of economic forecasts.
}
\examples{
# Generate sample data
y <- rnorm(100)
haty <- y + rnorm(100, sd = 0.5)
# Perform Mincer-Zarnowitz test
results <- MZ.test.mean(haty, y)
print(results)

}
\references{
Mincer, J. and Zarnowitz, V. (1969). The evaluation of economic forecasts. In Economic Forecasts and Expectations: Analysis of Forecasting Behavior and Performance, pages 3–46. National Bureau of Economic Research, Inc.
}
