% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_asyfuns.R
\name{MZ.test.quantile}
\alias{MZ.test.quantile}
\title{Mincer–Zarnowitz Quantile Regression Test}
\usage{
MZ.test.quantile(x, y, alpha)
}
\arguments{
\item{x}{Forecast values.}

\item{y}{Realized outcomes.}

\item{alpha}{Quantile level for the regression (e.g., 0.25).}
}
\value{
A \code{list} with:
\itemize{
  \item \code{estimate} – summary of the quantile regression fit.
  \item \code{WaldStat.MCB}, \code{Wald.pval.MCB} – test for \eqn{(\beta_0, \beta_1) = (0, 1)}.
  \item \code{WaldStat.DSC}, \code{Wald.pval.DSC} – test for \eqn{\beta_1 = 0}.
  \item \code{fittedvalues} – tibble with \code{x} and fitted \code{quantile regression} values.
}
}
\description{
Performs Mincer–Zarnowitz tests based on quantile regression \eqn{y_t = \beta_0 + \beta_1 x_t + u_t}.
Computes Wald statistics to test for zero MCB and/or zero DSC components:
\itemize{
  \item Null hypothesis \strong{MCB = 0:} joint test \eqn{(\beta_0, \beta_1) = (0, 1)} for mean (or median/quantile) calibration.
  \item Null hypothesis \strong{DSC = 0:} test \eqn{\beta_1 = 0} for discrimination skill.
}
The test for \strong{MCB = 0} is VQR backtest by Gaglianone et al (2011).
}
\examples{
set.seed(123)
y <- rnorm(100)
x <- y + rnorm(100)
MZ.test.quantile(x, y, alpha = 0.5)

}
\references{
Gaglianone, W. P., Lima, L. R., Linton, O., & Smith, D. R. (2011). Evaluating Value-at-Risk models via quantile regression. \emph{Journal of Business and Economic Statistics}, 29(1), 150–160.
\doi{10.1198/jbes.2010.08288}
}
\seealso{
\code{\link[quantreg]{rq}} for quantile regression estimation.
}
