% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{MZdec}
\alias{MZdec}
\title{Flexible MZ Mean Regression Using Different Scoring Functions}
\usage{
MZdec(x, y, S)
}
\arguments{
\item{x}{Numeric vector: Forecasts to be evaluated.}

\item{y}{Numeric vector: Realized values (must have the same length as x).}

\item{S}{function: Scoring function, which must be a function S(x, y).}
}
\value{
A list containing the following components:
  \item{x_rc}{Numeric vector: MZ recalibrated mean forecasts under the given scoring function.}
  \item{coef}{Numeric vector: Coefficients of the MZ regression under the given scoring function.}
  \item{dec}{tibble: Decomposition of the average score.}
}
\description{
This function fits a mean MZ regression using different loss functions S(x, y), e.g., for the squared error S(x, y) = (x - y)^2,
where the forecast is x and the realization is y. All scoring functions must be Bregman scores.
It returns the decomposition of the average score and the coefficients of the MZ regression under the given scoring function.
}
\details{
The function uses optimization techniques to estimate the parameters of the mean forecast model based on the specified loss function.
Users should ensure that the scoring function adheres to the properties of Bregman scores.
}
\examples{
x <- c(1, 2, 3, 4, 5)
y <- c(1.1, 2.2, 3.2, 4.1, 5.3)
result <- MZdec(
  x = x,
  y = y,
  S = function(x, y) (x - y)^2
)

}
\seealso{
\code{\link{optim}}, \code{\link{lm}}
}
\keyword{forecast}
\keyword{function}
\keyword{loss}
\keyword{mean}
\keyword{optimization}
