% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_asyfuns.R
\name{q_asy_var_dm}
\alias{q_asy_var_dm}
\title{Calculate variance of the decomposition components for two (rival) quantile forecasts based on Mincer-Zarnowitz regressions.}
\usage{
q_asy_var_dm(
  X_1,
  X_2,
  Y,
  S,
  V,
  alpha,
  vcov_estimator = sandwich::vcovHAC,
  tol = 50
)
}
\arguments{
\item{X_1}{Numeric vector  for the first forecast.}

\item{X_2}{Numeric vector for the second forecast.}

\item{Y}{Numeric vector for the outcome variable.}

\item{S}{Function, the scoring function used to assess the mean forecasts.}

\item{V}{Function, the identification function, where V is the derivative of S.}

\item{alpha}{Quantile level, \eqn{\alpha \in (0,1)}.}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}

\item{tol}{It can happen the determinat is not meaningfully calculated. We check for it with `det(Omega) <= 10^(-tol)`.
In this case the `Omega` output is a 5x5 matrix of `NaN` values.}
}
\value{
A list containing several components:
  - `asy_vars`: a data frame with the variances of s, mcb, and dsc for both forecasts.
  - `pvals`: p-values for testing the significance of differences.
  - `Lambda`: matrix used in the variance calculations.
  - `temp_Omega`: raw matrix of a_{1,t}, ..., e_t that is used to estimat cov matrix
  - `Omega`: covariance matrix based on the scoring function.
  - `eta`: transformation matrix used in variance calculations.
  - `Gamma`: product of eta and Lambda matrices.
  - `dec_1`: decomposition results for the first forecast.
  - `dec_2`: decomposition results for the second forecast.
}
\description{
This function computes variance of the score decomposition for two rival quantile forecasts X_1 and X_2
based on the provided data. It calculates components related to the score (s), miscalibration (mcb), and discrimination (dsc).
}
\details{
It can happen that covariance matrix estimation fails.
In this cases the `Omega` output is a 5x5 matrix of `NaN` values.
The p-values are set to 1 in this case to facilitate conservative test decisions.
}
\examples{
# tbc

}
