% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_asyfuns.R
\name{qdecomposition}
\alias{qdecomposition}
\title{Decomposition of Quantile Scores}
\usage{
qdecomposition(x, x_rc, y, S, alpha)
}
\arguments{
\item{x}{Numeric vector of forecast values.}

\item{x_rc}{Numeric vector of recalibrated forecast values.}

\item{y}{Numeric vector of actual outcome values.}

\item{S}{Function, a user-defined generalized piecewise linear (GPL) scoring function that takes thre arguments (forecast, outcome, alpha)
and returns a numeric value indicating the score.}

\item{alpha}{Quantile level.}
}
\value{
A tibble with columns:
  - `s`: the average score using the original forecast.
  - `mcb`: miscalibration component
  - `dsc`: discrimination component
  - `unc`: uncertainty component
}
\description{
This function performs a decomposition of the average score given forecasts,
recalibrated forecasts, and actual outcomes. It allows for flexible use of any quantile scoring function `S`
provided by the user, to calculate the mean score, marginal contribution bias (MCB),
and the decomposition of the scoring contribution (DSC).
}
\examples{
# tbc

}
