% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_asyfuns.R
\name{qMZdec}
\alias{qMZdec}
\title{Flexible MZ-style Quantile Regression Using Different GPL Scoring Functions}
\usage{
qMZdec(x, y, alpha, S)
}
\arguments{
\item{x}{Numeric vector: Forecasts to be evaluated.}

\item{y}{Numeric vector: Realized values (must have the same length as x).}

\item{alpha}{Quantile level, \eqn{\alpha \in (0,1)}.}

\item{S}{function: Scoring function, which must be a function S(x, y, alpha).}
}
\value{
A list containing the following components:
  \item{x_rc}{Numeric vector: MZ recalibrated mean forecasts under the given scoring function.}
  \item{coef}{Numeric vector: Coefficients of the MZ regression under the given scoring function.}
  \item{dec}{tibble: Decomposition of the average score.}
}
\description{
This function fits a quantile MZ-style regression using different generalized piecewise linear (GPL) scoring functions of order alpha S(x, y, alpha), e.g., for the asymmetric piecewise linear scoring function \eqn{S(x,y) = (\mathbb{1}\{y < x\} - \alpha)(x-y)}.,
where the forecast is x and the realization is y. All scoring functions must be Bregman scores.
It returns the decomposition of the average score and the coefficients of the MZ regression under the given scoring function.
}
\details{
The function uses optimization techniques to estimate the parameters of the mean forecast model based on the specified loss function.
Users should ensure that the scoring function adheres to the properties of Bregman scores.
}
\examples{
# tbc

}
\seealso{
\code{\link{optim}}, \code{\link{lm}}
}
\keyword{forecast}
\keyword{function}
\keyword{loss}
\keyword{mean}
\keyword{optimization}
