% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi.R
\name{qSDI}
\alias{qSDI}
\title{Score Decomposition Inference (SDI) for Quantile Forecasts}
\usage{
qSDI(X_1, X_2, Y, S, V, alpha, vcov_estimator = sandwich::vcovHAC, tol = 50)
}
\arguments{
\item{X_1}{Numeric vector for the first forecast.}

\item{X_2}{Numeric vector for the second forecast.}

\item{Y}{Numeric vector for the outcome variable.}

\item{S}{Function, the scoring function used to assess the quantile forecasts.}

\item{V}{Function, the identification function, where V is the derivative of S.}

\item{alpha}{Quantile level, \eqn{\alpha \in (0,1)}.}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}

\item{tol}{See `asy_var_dm()`.}
}
\value{
A list of class "SDI" containing several components:
  - `asyvardm`: a list with the variance decomposition results including `asy_vars`, `pvals`, `Lambda`, `Omega`, `eta`, `Gamma`, `dec_1`, and `dec_2`.
  - `mcbnulltest_X1`: results of the miscalibration null test for the first forecast.
  - `mcbnulltest_X2`: results of the miscalibration null test for the second forecast.
  - `mcbIUUI`: combined p-values for the miscalibration tests.
  - `dscnulltest_X1`: placeholder for discrimination null test for the first forecast.
  - `dscnulltest_X2`: placeholder for discrimination null test for the second forecast.
  - `dscIUUI`: placeholder for combined p-values for the discrimination test.
}
\description{
This function provides a wrapper for calculating the variance of the decomposition components for two (rival) mean forecasts
based on Mincer-Zarnowitz style quantile regressions, along with performing null tests for miscalibration and discrimination.
}
\examples{
# tbc

}
