% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi.R
\name{SDI}
\alias{SDI}
\title{Score Decomposition Inference (SDI)}
\usage{
SDI(
  X_1,
  X_2,
  Y,
  S,
  V,
  Spp = NA,
  selectionvector = matrix(c(1, 1, 1, 1), nrow = 1, ncol = 4),
  vcov_estimator = sandwich::vcovHAC
)
}
\arguments{
\item{X_1}{Numeric vector for the first forecast.}

\item{X_2}{Numeric vector for the second forecast.}

\item{Y}{Numeric vector for the outcome variable.}

\item{S}{Function, the scoring function used to assess the forecasts.}

\item{V}{Function, the identification function, where V is the derivative of S.}

\item{Spp}{Function, second derivative of the scoring function, default is NA.}

\item{selectionvector}{A (optional) 1x4 matrix object specifying the selection vector. Default is a 1x4 matrix of ones.}

\item{vcov_estimator}{Function, variance-covariance estimator from the sandwich package.}
}
\value{
A list of class "SDI" containing several components:
  - `asyvardm`: a list with the variance decomposition results including `asy_vars`, `pvals`, `Lambda`, `Omega`, `eta`, `Gamma`, `dec_1`, and `dec_2`.
  - `mcbnulltest_X1`: results of the miscalibration null test for the first forecast.
  - `mcbnulltest_X2`: results of the miscalibration null test for the second forecast.
  - `mcbIUUI`: combined p-values for the miscalibration tests.
  - `dscnulltestX1`: placeholder for discrimination null test for the first forecast.
  - `dscnulltestX2`: placeholder for discrimination null test for the second forecast.
  - `dscIUUI`: placeholder for combined p-values for the discrimination test.
  - `mySDIvar`: the selected variance of the SDI.
  - `dm`: a data frame with the Diebold-Mariano test results including `dm_asy_var`, `dm_stat`, and `dm_pval`.
}
\description{
This function provides a wrapper for calculating the variance of the decomposition components for two (rival) forecasts
based on Mincer-Zarnowitz regressions, along with performing null tests for miscalibration and discrimination.
}
\examples{
# Define a simple scoring function
S <- function(y, x) { (y - x)^2 }
# Define its derivative(s)
V <- function(x, y) { 2 * (y - x) }
Spp <- function(x, y) { 2 }
# Generate sample data
Y <- rnorm(100)
X_1 <- Y + rnorm(100, sd = 0.5)
X_2 <- Y + rnorm(100, sd = 0.5)
# Calculate score decomposition inference
results <- SDI(X_1, X_2, Y, S, V, Spp)
print(results)

}
