% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyfuns.R
\name{decomposition}
\alias{decomposition}
\title{Decomposition of the Bregman Score}
\usage{
decomposition(x, x_rc, y, S)
}
\arguments{
\item{x}{Numeric vector of forecast values.}

\item{x_rc}{Numeric vector of recalibrated forecast values.}

\item{y}{Numeric vector of actual outcome values.}

\item{S}{Function, a user-defined scoring function that takes two arguments (forecast, outcome)
and returns a numeric value indicating the score.}
}
\value{
A tibble with columns:
  - `s`: the average score using the original forecast.
  - `mcb`: miscalibration component
  - `dsc`: discrimination component
  - `unc`: uncertainty component
}
\description{
This function performs a decomposition of any Bregman scoring function given forecasts,
recalibrated forecasts, and actual outcomes. It allows for flexible use of any Bregman scoring function `S`
provided by the user, to calculate the average score, the miscalibration component (MCB), the
discrimination component (DSC), and the uncertainty component.
}
\examples{
# Define a simple scoring function
S <- function(x, y) { (y - x)^2 }
# Generate sample data
y <- rnorm(100)
x <- y + rnorm(100, sd = 0.5)
x_rc <- y + rnorm(100, sd = 0.2)
# Decompose the scoring
decomposition_results <- decomposition(x, x_rc, y, S)
print(decomposition_results)

}
